+++ packages/orq-rc/RELEASES.md
@@ -6669,3 +6669,13 @@ Based on:
 - [typescript v4.3.0-rc.53] packages/orq-rc
 ### Releases
 - [NPM v4.3.0-rc.53] https://www.npmjs.com/package/@orq-ai/node/v/4.3.0-rc.53 - packages/orq-rc
+
+## 2026-02-11 13:18:57
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.710.0 (2.818.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.3.0-rc.54] packages/orq-rc
+### Releases
+- [NPM v4.3.0-rc.54] https://www.npmjs.com/package/@orq-ai/node/v/4.3.0-rc.54 - packages/orq-rc